cleaning up main.cpp
