+++ prj.conf
@@ -28,6 +28,7 @@ CONFIG_ZBUS_LOG_LEVEL_INF=y
 CONFIG_ZBUS_CHANNEL_NAME=y
 CONFIG_ZBUS_OBSERVER_NAME=y
 
-# Bootloader
-CONFIG_BOOTLOADER_MCUBOOT=y
-CONFIG_MCUBOOT_SIGNATURE_KEY_FILE="deps/bootloader/mcuboot/root-rsa-2048.pem"
+# # Bootloader
+# CONFIG_BOOTLOADER_MCUBOOT=y
+# CONFIG_MCUBOOT_GENERATE_UNSIGNED_IMAGE=n
+# CONFIG_MCUBOOT_SIGNATURE_KEY_FILE="deps/bootloader/mcuboot/root-rsa-2048.pem"
