+++ src/AppSensorDataProducer.cpp
@@ -6,9 +6,6 @@
 // User C++ class headers
 #include "EventManager.h"
 
-// Import channel from another file
-ZBUS_CHAN_DECLARE(eventsChannel);
-
 // Function declaration of thread handlers
 static void sensorDataProducerThreadHandler();
 
@@ -57,11 +54,8 @@ static void sensorDataProducerThreadHandler() {
               printk("Started acquiring sensor data and saving it to storage");
 
               // Fake acquisition (to be replaced with real sensor acquisition function)
-              for (size_t i = 0; i < 10; i++) {
-                printk(".");
-                k_msleep(500);
-              }
-              printk("\r\n");
+              k_msleep(5000);
+              printk("\r\n===========================================================\r\n");
 
               // Publish the <EVENT_SENSOR_DATA_SAVED> event on <eventsChannel>
               event.id = EVENT_SENSOR_DATA_SAVED;
